fix: don't flicker root DropZone when dragging
There's a slight delay between the user starting the drag and draggedItem being set, which meant that the root dropzone was being enabled too late after the user would start dragging. This fixes that by using an `onMouseDown` and `onMouseUp` event handler to introduce a new `userWillDrag` state within the current context.