perf: improve render performance of fields
Running the dispatcher on each change is expensive, and waiting for it to run before showing the user's input creates a significant lag.

This change uses local state for all inputs, to give the user immediate feedback, and debounces the dispatcher call.

useDeferredValue doesn't help us here, since the reducer is expensive regardless of the UI render.