fix: don't reset array item labels when changing order
When the user is relying on the default label (i.e. Item 1) for items in an array field, the index would reset whenever the user dragged the items between positions. That was because the label was based on the index.

This fix stores the original index for the item in the arrayState, which persists between drags.

This also introduces some refactors that may improve the general robustness of array fields.