refactor: enable fields to set UI state when changing values
This means that UI state and the data is changed at the same time, resulting in a single history entry and stopping race conditions for more advanced fields that may need to set UI state alongside the value. For example, the array field type creates a mirror of the UI state. Rewinding and fast-forwarding history historically meant that the UI mirror would become out of sync with the data value. Oftentimes, this would go unnoticed, but it can corrupt the data in subtle ways.

Additionally, this change will also be available to custom fields.