refactor: don't record zoom in UI state
It doesn't make sense for this to be recorded in UI state as it changes dynamically based on the user window size. Additionally, the initial zoom value was not getting correctly recorded in the initial history value as the zoom value cannot be determined until the page has rendered, which is after the initial history value is set.