perf: don't lock main thread when iframe styles changed
This may occur in dev environments when styles are injected into the parent <head>, causing issues if this occurs during a drop. For example, using @mui/material in form inputs injects styles as soon as an item is dropped via emotion, causing the main thread to hang. Deferring this until after the main thread is completed via measured/auto-frame-component@0.1.3 addresses the issue.