feat: add APIs for setting undo/redo history
This addresses https://github.com/measuredco/puck/issues/519.

This change exposes the history store's setHistories and setHistoryIndex
functions. When they are called the data is dispatched so that Puck
rerenders. The use case this supports is when a Puck consumer wants to
change or reset the history state manually without requiring the page to
be refreshed.

---------

Co-authored-by: Chris Villa <chrisvxd@users.noreply.github.com>