refactor: clear deepest droppable ref on drag end for touch
When using pointer, the active zone is constantly recalculated due to pointer movement. For touch, the active zone is only calculated during drag. Certain scenarios could cause the deepest droppable to change, but not update the state as the deepest droppable ref is still the same. Resetting this after each drag solves this.